카테고리 베스트 아이템 버그 수정 및 카테고리 아이템 쿼리 성능 최적화

+++ src/main/java/org/example/tamaapi/dto/responseDto/CustomPage.java
@@ -23,7 +23,7 @@ public class CustomPage<T> {
     }
 
     //직접 만든 페이징.
-    public CustomPage(List<T> content, CustomPageRequest customPageRequest, int rowCount) {
+    public CustomPage(List<T> content, CustomPageRequest customPageRequest, Long rowCount) {
         this.content = content;
         myPageable = new CustomPageable(customPageRequest.getPage(), customPageRequest.getSize(), rowCount);
     }
